backend/requirements.txt
# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
websockets>=12.0
aiofiles>=23.2.0

# Database
sqlalchemy[asyncio]>=2.0.30
aiosqlite>=0.20.0

# OCR
easyocr>=1.7.0

# Image processing
opencv-python-headless>=4.9.0
Pillow>=10.3.0
simple-lama-inpainting>=0.1.0
numpy>=1.26.0

# AI
anthropic>=0.40.0

# MCP
mcp[cli]>=1.0.0

# Config
pydantic>=2.7.0
pydantic-settings>=2.3.0
python-dotenv>=1.0.0

# HTTP client (for connectors)
httpx>=0.27.0
beautifulsoup4>=4.12.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
